requirements.txt
torch~=1.12.1+cu116
opencv-python~=4.7.0.68
pandas~=1.5.1
numpy~=1.23.3
torchvision~=0.13.1+cu116
Pillow~=9.2.0